refactor(trusty-gworkspace,trusty-memory): split 2 files under 500-SLOC cap (#1195)
Behavior-preserving refactor of the two remaining over-cap production
files in these crates (#1195):

- trusty-gworkspace/src/tools.rs (557 SLOC) -> tools/ module: thin
  tools/mod.rs dispatcher + per-service submodules (schema, accounts,
  calendar, gmail, drive, docs, sheets, slides, tasks). Public path
  `crate::tools::tool_list_response` is unchanged.
- trusty-memory/src/lib.rs (623 SLOC -> 443) -> extracted the daemon
  event vocabulary into events.rs (DaemonEvent, HookType, InjectionKind,
  open_activity_log_with_fallback) and the HTTP/SSE serving surface into
  http_server.rs (run_http*, bind_dynamic_port, address-file helpers,
  sse_handler). Both are re-exported at the crate root so the public API
  (trusty_memory::DaemonEvent, trusty_memory::run_http_on, ...) is
  byte-for-byte unchanged.

Drops both files from .line-cap-allowlist.tsv (ratchet-down).
