+++ .line-cap-allowlist.tsv
@@ -7,5 +7,3 @@
 # and /* ... */ block comments (including multi-line spans). Trailing-comment lines count.
 # Ratchet: budgets may only DECREASE; when a file drops <= its applicable cap, remove it.
 # Regenerate with: scripts/check_line_cap.sh --update  (use --seed only to bootstrap).
-crates/trusty-gworkspace/src/tools.rs	557
-crates/trusty-memory/src/lib.rs	618
